requirements.txt
aiosqlite==0.21.0
fastapi==0.124.4
fastmcp==2.14.0
langchain_core==1.2.0
langchain_mcp_adapters==0.2.1
langchain_ollama==1.0.1
langchain_openai==1.1.3
langgraph==1.0.5
pydantic==2.12.5
python-dotenv==1.2.1
rich==14.2.0
